Rename filenames according to the guidelines
