+++ apps/frontend/src/app/5_pages/App/App.module.css
@@ -1,0 +1,9 @@
+.test {
+  @apply font-extrabold;
+
+  color: orange;
+
+  &:hover {
+    color: blue;
+  }
+}
