+++ apps/frontend/src/app/5_pages/App/App.test.tsx
@@ -1,5 +1,7 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
+
+import React from 'react';
+
 import App from './App';
 
 test('renders learn react link', () => {
