--- apps/frontend/src/app/5_pages/index.module.css
@@ -1,9 +1,0 @@
-.test {
-    @apply font-extrabold;
-
-    color: orange;
-
-    &:hover{
-        color: blue;
-    };
-}
