+++ apps/frontend/src/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
+
 import ReactDOM from 'react-dom/client';
-import App from './app/5_pages/App';
+
+import App from './app/5_pages/App/App';
 import './styles/tailwindcss/index.css';
 
 const root = ReactDOM.createRoot(
